pyproject.toml
[tool.poetry]
name = "sdss-yao"
version = "0.1.0-alpha.0"
description = "BOSS spectrograph actor that uses an STA Archon controller"
authors = ["José Sánchez-Gallego <gallegoj@uw.edu>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/yao"
repository = "https://github.com/sdss/yao"
documentation = "https://sdss-yao.readthedocs.org"
keywords = ["astronomy", "software"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
	"Programming Language :: Python :: 3.6",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "yao" }
]
include = ["yao/etc/*"]

[tool.poetry.build]
script = "build.py"
generate-setup-file = false

[tool.poetry.scripts]
yao = "yao.__main__:yao"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
sdss-archon = {path = "../archon_boss", develop = true}
sdss-tree = ">=2.15.2"
sdss-access = ">=0.2.3"
sdsstools = ">=0.4.0"
click = "^8.1.2"

[tool.poetry.dev-dependencies]
ipython = ">=8.0.0"
flake8 = ">=3.7.9"
doc8 = ">=0.8.0"
pytest = ">=5.2.2"
pytest-asyncio = ">=0.10.0"
pytest-cov = ">=2.8.1"
pytest-mock = ">=1.13.0"
pytest-sugar = ">=0.9.2"
isort = ">=4.3.21"
codecov = ">=2.0.15"
coverage = {version = ">=5.0", extras = ["toml"]}
ipdb = ">=0.12.3"
rstcheck = ">=3.3.1"
Sphinx = ">=3.0.0"
black = ">=20.8b1"
jedi = ">=0.17.2"
sphinx-click = ">=2.6.0"
sphinx-jsonschema = ">=1.16.7"
myst-parser = ">=0.14.0"
furo = ">=2021.6.18-beta.36"
nox = ">=2021.6.12"
sphinx-autobuild = ">=2021.3.14"
sphinx-copybutton = ">=0.3.3"

[tool.black]
line-length = 88
target-version = ['py39']
fast = true

[tool.isort]
profile = "black"
sections = ["FUTURE", "STDLIB", "TYPING", "THIRDPARTY", "SDSS", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = "yao"
known_typing = ["typing"]
known_sdss = ["sdsstools", "archon"]
lines_after_imports = 2

[tool.pytest.ini_options]
addopts = "--cov yao --cov-report xml --cov-report html --cov-report term"

[tool.coverage.run]
branch = true
include = ["yao/*"]
omit = [
    "*/__init__.py",
    "yao/__main__.py"
]

[tool.coverage.report]
exclude_lines = [
    "# pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:"
]

[build-system]
requires = ["poetry-core>=1.1.0a7"]
build-backend = "poetry.core.masonry.api"
